split get contact flow
