+++ solutions/Virtual-Assistant/src/csharp/skills/calendarskill/calendarskill/Dialogs/ConfirmRecipient/ConfirmRecipientDialog.cs
@@ -1,0 +1,431 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Collections.Specialized;
+using System.Linq;
+using System.Threading;
+using System.Threading.Tasks;
+using CalendarSkill.Dialogs.CreateEvent.Resources;
+using CalendarSkill.Dialogs.Shared;
+using CalendarSkill.Dialogs.Shared.DialogOptions;
+using CalendarSkill.Models;
+using CalendarSkill.ServiceClients;
+using Luis;
+using Microsoft.Bot.Builder;
+using Microsoft.Bot.Builder.Dialogs;
+using Microsoft.Bot.Builder.Dialogs.Choices;
+using Microsoft.Bot.Solutions.Extensions;
+using Microsoft.Bot.Solutions.Skills;
+using Microsoft.Bot.Solutions.Util;
+
+namespace CalendarSkill.Dialogs.ConfirmRecipient
+{
+    public class ConfirmRecipientDialog : CalendarSkillDialog
+    {
+        public ConfirmRecipientDialog(
+            SkillConfigurationBase services,
+            IStatePropertyAccessor<CalendarSkillState> accessor,
+            IServiceManager serviceManager,
+            IBotTelemetryClient telemetryClient)
+            : base(nameof(ConfirmRecipientDialog), services, accessor, serviceManager, telemetryClient)
+        {
+            TelemetryClient = telemetryClient;
+
+            var updateAddress = new WaterfallStep[]
+            {
+                UpdateAddress,
+                AfterUpdateAddress,
+            };
+
+            var confirmAttendee = new WaterfallStep[]
+            {
+                ConfirmAttendee,
+                AfterConfirmAttendee,
+            };
+
+            var updateName = new WaterfallStep[]
+            {
+                UpdateUserName,
+                AfterUpdateUserName,
+            };
+
+            // Define the conversation flow using a waterfall model.
+            AddDialog(new WaterfallDialog(Actions.UpdateAddress, updateAddress) { TelemetryClient = telemetryClient });
+            AddDialog(new WaterfallDialog(Actions.ConfirmAttendee, confirmAttendee) { TelemetryClient = telemetryClient });
+            AddDialog(new WaterfallDialog(Actions.UpdateName, updateName) { TelemetryClient = telemetryClient });
+
+            InitialDialogId = Actions.UpdateAddress;
+        }
+
+        // update address waterfall steps
+        public async Task<DialogTurnResult> UpdateAddress(WaterfallStepContext sc, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            try
+            {
+                var state = await Accessor.GetAsync(sc.Context, cancellationToken: cancellationToken);
+                if (state.AttendeesNameList.Any())
+                {
+                    return await sc.NextAsync(cancellationToken: cancellationToken);
+                }
+
+                return await sc.PromptAsync(Actions.Prompt, new PromptOptions { Prompt = sc.Context.Activity.CreateReply(CreateEventResponses.NoAttendees) }, cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                await HandleDialogExceptions(sc, ex);
+                return new DialogTurnResult(DialogTurnStatus.Cancelled, CommonUtil.DialogTurnResultCancelAllDialogs);
+            }
+        }
+
+        public async Task<DialogTurnResult> AfterUpdateAddress(WaterfallStepContext sc, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            try
+            {
+                var state = await Accessor.GetAsync(sc.Context, cancellationToken: cancellationToken);
+                if (state.AttendeesNameList.Any())
+                {
+                    return await sc.BeginDialogAsync(Actions.ConfirmAttendee, cancellationToken: cancellationToken);
+                }
+
+                if (sc.Result != null)
+                {
+                    sc.Context.Activity.Properties.TryGetValue("OriginText", out var content);
+                    var userInput = content != null ? content.ToString() : sc.Context.Activity.Text;
+
+                    // TODO: can we do this somewhere else
+                    if (IsEmail(userInput))
+                    {
+                        state.Attendees.Add(new EventModel.Attendee { Address = userInput });
+                        return await sc.EndDialogAsync(true, cancellationToken);
+                    }
+                    else
+                    {
+                        if (state.EventSource != EventSource.Other)
+                        {
+                            if (userInput != null)
+                            {
+                                var nameList = userInput.Split(CreateEventWhiteList.GetContactNameSeparator(), StringSplitOptions.None)
+                                    .Select(x => x.Trim())
+                                    .Where(x => !string.IsNullOrWhiteSpace(x))
+                                    .ToList();
+                                state.AttendeesNameList = nameList;
+                            }
+
+                            return await sc.BeginDialogAsync(Actions.ConfirmAttendee, cancellationToken: cancellationToken);
+                        }
+                        else
+                        {
+                            return await sc.BeginDialogAsync(Actions.UpdateAddress, new UpdateAddressDialogOptions(UpdateAddressDialogOptions.UpdateReason.NotAnAddress), cancellationToken);
+                        }
+                    }
+                }
+                else
+                {
+                    return await sc.NextAsync(cancellationToken: cancellationToken);
+                }
+            }
+            catch (Exception ex)
+            {
+                await HandleDialogExceptions(sc, ex);
+                return new DialogTurnResult(DialogTurnStatus.Cancelled, CommonUtil.DialogTurnResultCancelAllDialogs);
+            }
+        }
+
+        // confirm attendee waterfall steps
+        public async Task<DialogTurnResult> ConfirmAttendee(WaterfallStepContext sc, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            try
+            {
+                var state = await Accessor.GetAsync(sc.Context, cancellationToken: cancellationToken);
+                var currentRecipientName = state.AttendeesNameList[state.ConfirmAttendeesNameIndex];
+                if (IsEmail(currentRecipientName))
+                {
+                    var result =
+                        new FoundChoice()
+                        {
+                            Value = $"{currentRecipientName}: {currentRecipientName}",
+                        };
+
+                    return await sc.NextAsync(result);
+                }
+
+                var originPersonList = await GetPeopleWorkWithAsync(sc, currentRecipientName);
+                var originContactList = await GetContactsAsync(sc, currentRecipientName);
+                originPersonList.AddRange(originContactList);
+
+                var originUserList = new List<PersonModel>();
+                try
+                {
+                    originUserList = await GetUserAsync(sc, currentRecipientName);
+                }
+                catch
+                {
+                    // do nothing when get user failed. because can not use token to ensure user use a work account.
+                }
+
+                (var personList, var userList) = FormatRecipientList(originPersonList, originUserList);
+
+                // todo: should set updatename reason in sc.Result
+                if (personList.Count > 10)
+                {
+                    return await sc.BeginDialogAsync(Actions.UpdateName, new UpdateUserNameDialogOptions(UpdateUserNameDialogOptions.UpdateReason.TooMany), cancellationToken);
+                }
+
+                if (personList.Count < 1 && userList.Count < 1)
+                {
+                    return await sc.BeginDialogAsync(Actions.UpdateName, new UpdateUserNameDialogOptions(UpdateUserNameDialogOptions.UpdateReason.NotFound), cancellationToken);
+                }
+
+                if (personList.Count == 1)
+                {
+                    var user = personList.FirstOrDefault();
+                    if (user != null)
+                    {
+                        var result =
+                            new FoundChoice()
+                            {
+                                Value = $"{user.DisplayName}: {user.Emails[0] ?? user.UserPrincipalName}",
+                            };
+
+                        return await sc.NextAsync(result, cancellationToken);
+                    }
+                }
+
+                // TODO: should be simplify
+                var selectOption = await GenerateOptions(personList, userList, sc);
+
+                // If no more recipient to show, start update name flow and reset the recipient paging index.
+                if (selectOption.Choices.Count == 0)
+                {
+                    state.ShowAttendeesIndex = 0;
+                    return await sc.BeginDialogAsync(Actions.UpdateName, new UpdateUserNameDialogOptions(UpdateUserNameDialogOptions.UpdateReason.NotFound), cancellationToken);
+                }
+
+                // Update prompt string to include the choices because the list style is none;
+                // TODO: should be removed if use adaptive card show choices.
+                var choiceString = GetSelectPromptString(selectOption, true);
+                selectOption.Prompt.Text = choiceString;
+                return await sc.PromptAsync(Actions.Choice, selectOption, cancellationToken);
+            }
+            catch (SkillException ex)
+            {
+                await HandleDialogExceptions(sc, ex);
+                return new DialogTurnResult(DialogTurnStatus.Cancelled, CommonUtil.DialogTurnResultCancelAllDialogs);
+            }
+            catch (Exception ex)
+            {
+                await HandleDialogExceptions(sc, ex);
+                return new DialogTurnResult(DialogTurnStatus.Cancelled, CommonUtil.DialogTurnResultCancelAllDialogs);
+            }
+        }
+
+        public async Task<DialogTurnResult> AfterConfirmAttendee(WaterfallStepContext sc, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            try
+            {
+                var state = await Accessor.GetAsync(sc.Context, cancellationToken: cancellationToken);
+
+                // result is null when just update the recipient name. show recipients page should be reset.
+                if (sc.Result == null)
+                {
+                    state.ShowAttendeesIndex = 0;
+                    return await sc.BeginDialogAsync(Actions.ConfirmAttendee, cancellationToken: cancellationToken);
+                }
+                else if (sc.Result.ToString() == General.Intent.Next.ToString())
+                {
+                    state.ShowAttendeesIndex++;
+                    return await sc.BeginDialogAsync(Actions.ConfirmAttendee, cancellationToken: cancellationToken);
+                }
+                else if (sc.Result.ToString() == General.Intent.Previous.ToString())
+                {
+                    if (state.ShowAttendeesIndex > 0)
+                    {
+                        state.ShowAttendeesIndex--;
+                    }
+
+                    return await sc.BeginDialogAsync(Actions.ConfirmAttendee, cancellationToken: cancellationToken);
+                }
+                else
+                {
+                    var user = (sc.Result as FoundChoice)?.Value.Trim('*');
+                    if (user != null)
+                    {
+                        var attendee = new EventModel.Attendee
+                        {
+                            DisplayName = user.Split(": ")[0],
+                            Address = user.Split(": ")[1],
+                        };
+                        if (state.Attendees.All(r => r.Address != attendee.Address))
+                        {
+                            state.Attendees.Add(attendee);
+                        }
+                    }
+
+                    state.ConfirmAttendeesNameIndex++;
+                    if (state.ConfirmAttendeesNameIndex < state.AttendeesNameList.Count)
+                    {
+                        return await sc.BeginDialogAsync(Actions.ConfirmAttendee, cancellationToken: cancellationToken);
+                    }
+
+                    return await sc.EndDialogAsync(true, cancellationToken);
+                }
+            }
+            catch (Exception ex)
+            {
+                await HandleDialogExceptions(sc, ex);
+                return new DialogTurnResult(DialogTurnStatus.Cancelled, CommonUtil.DialogTurnResultCancelAllDialogs);
+            }
+        }
+
+        // update name waterfall steps
+        public async Task<DialogTurnResult> UpdateUserName(WaterfallStepContext sc, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            try
+            {
+                var state = await Accessor.GetAsync(sc.Context, cancellationToken: cancellationToken);
+                var currentRecipientName = state.AttendeesNameList[state.ConfirmAttendeesNameIndex];
+
+                if (((UpdateUserNameDialogOptions)sc.Options).Reason == UpdateUserNameDialogOptions.UpdateReason.TooMany)
+                {
+                    return await sc.PromptAsync(Actions.Prompt, new PromptOptions { Prompt = sc.Context.Activity.CreateReply(CreateEventResponses.PromptTooManyPeople, ResponseBuilder, new StringDictionary() { { "UserName", currentRecipientName } }) }, cancellationToken);
+                }
+                else
+                {
+                    return await sc.PromptAsync(Actions.Prompt, new PromptOptions { Prompt = sc.Context.Activity.CreateReply(CreateEventResponses.PromptPersonNotFound, ResponseBuilder, new StringDictionary() { { "UserName", currentRecipientName } }) }, cancellationToken);
+                }
+            }
+            catch (Exception ex)
+            {
+                await HandleDialogExceptions(sc, ex);
+                return new DialogTurnResult(DialogTurnStatus.Cancelled, CommonUtil.DialogTurnResultCancelAllDialogs);
+            }
+        }
+
+        public async Task<DialogTurnResult> AfterUpdateUserName(WaterfallStepContext sc, CancellationToken cancellationToken = default(CancellationToken))
+        {
+            try
+            {
+                sc.Context.Activity.Properties.TryGetValue("OriginText", out var content);
+                var userInput = content != null ? content.ToString() : sc.Context.Activity.Text;
+                if (!string.IsNullOrEmpty(userInput))
+                {
+                    var state = await Accessor.GetAsync(sc.Context, cancellationToken: cancellationToken);
+                    state.AttendeesNameList[state.ConfirmAttendeesNameIndex] = userInput;
+                }
+
+                return await sc.EndDialogAsync(null, cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                await HandleDialogExceptions(sc, ex);
+                return new DialogTurnResult(DialogTurnStatus.Cancelled, CommonUtil.DialogTurnResultCancelAllDialogs);
+            }
+        }
+
+        protected async Task<PromptOptions> GenerateOptions(List<PersonModel> personList, List<PersonModel> userList, DialogContext dc)
+        {
+            var state = await Accessor.GetAsync(dc.Context);
+            var pageIndex = state.ShowAttendeesIndex;
+            var pageSize = 5;
+            var skip = pageSize * pageIndex;
+            var options = new PromptOptions
+            {
+                Choices = new List<Choice>(),
+                Prompt = dc.Context.Activity.CreateReply(CreateEventResponses.ConfirmRecipient),
+            };
+            for (var i = 0; i < personList.Count; i++)
+            {
+                var user = personList[i];
+                var mailAddress = user.Emails[0] ?? user.UserPrincipalName;
+
+                var choice = new Choice()
+                {
+                    Value = $"**{user.DisplayName}: {mailAddress}**",
+                    Synonyms = new List<string> { (options.Choices.Count + 1).ToString(), user.DisplayName, user.DisplayName.ToLower(), mailAddress },
+                };
+
+                var userName = user.UserPrincipalName?.Split("@").FirstOrDefault() ?? user.UserPrincipalName;
+                if (!string.IsNullOrEmpty(userName))
+                {
+                    choice.Synonyms.Add(userName);
+                    choice.Synonyms.Add(userName.ToLower());
+                }
+
+                if (skip <= 0)
+                {
+                    if (options.Choices.Count >= pageSize)
+                    {
+                        return options;
+                    }
+
+                    options.Choices.Add(choice);
+                }
+                else
+                {
+                    skip--;
+                }
+            }
+
+            if (options.Choices.Count == 0)
+            {
+                pageSize = 10;
+            }
+
+            for (var i = 0; i < userList.Count; i++)
+            {
+                var user = userList[i];
+                var mailAddress = user.Emails[0] ?? user.UserPrincipalName;
+                var choice = new Choice()
+                {
+                    Value = $"{user.DisplayName}: {mailAddress}",
+                    Synonyms = new List<string> { (options.Choices.Count + 1).ToString(), user.DisplayName, user.DisplayName.ToLower(), mailAddress },
+                };
+
+                var userName = user.UserPrincipalName?.Split("@").FirstOrDefault() ?? user.UserPrincipalName;
+                if (!string.IsNullOrEmpty(userName))
+                {
+                    choice.Synonyms.Add(userName);
+                    choice.Synonyms.Add(userName.ToLower());
+                }
+
+                if (skip <= 0)
+                {
+                    if (options.Choices.Count >= pageSize)
+                    {
+                        return options;
+                    }
+
+                    options.Choices.Add(choice);
+                }
+                else if (skip >= 10)
+                {
+                    return options;
+                }
+                else
+                {
+                    skip--;
+                }
+            }
+
+            return options;
+        }
+
+        private string GetSelectPromptString(PromptOptions selectOption, bool containNumbers)
+        {
+            var result = string.Empty;
+            result += selectOption.Prompt.Text + "\r\n";
+            for (var i = 0; i < selectOption.Choices.Count; i++)
+            {
+                var choice = selectOption.Choices[i];
+                result += "  ";
+                if (containNumbers)
+                {
+                    result += (i + 1) + "-";
+                }
+
+                result += choice.Value + "\r\n";
+            }
+
+            return result;
+        }
+    }
+}
