+++ solutions/Virtual-Assistant/src/csharp/skills/calendarskill/calendarskill/Dialogs/ConfirmRecipient/Resources/ConfirmRecipientResponses.tt
@@ -1,0 +1,3 @@
+﻿<#@ template debug="false" hostspecific="true" language="C#" #>
+<#@ output extension=".cs" #>
+<#@ include file="..\..\..\..\..\..\Microsoft.Bot.Solutions\Resources\ResponsesTemplate.t4"#>
