+++ solutions/Virtual-Assistant/src/csharp/skills/calendarskill/calendarskill/Dialogs/Shared/Actions.cs
@@ -35,5 +35,6 @@
         public const string Read = "read";
         public const string Greeting = "greeting";
         public const string GetEventsInit = "getEventsInit";
+        public const string ConfirmRecipientDialog = "ConfirmRecipientDialog";
     }
 }
