+++ solutions/Virtual-Assistant/src/csharp/skills/calendarskill/calendarskill/Dialogs/Shared/DialogOptions/CalendarSkillDialogOptions.cs
@@ -1,4 +1,4 @@
-﻿namespace CalendarSkill
+﻿namespace CalendarSkill.Dialogs.Shared.DialogOptions
 {
     public class CalendarSkillDialogOptions
     {
