+++ solutions/Virtual-Assistant/src/csharp/skills/calendarskill/calendarskill/Dialogs/Shared/DialogOptions/UpdateAddressDialogOptions.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license.
 
-namespace CalendarSkill
+namespace CalendarSkill.Dialogs.Shared.DialogOptions
 {
     public class UpdateAddressDialogOptions
     {
