+++ solutions/Virtual-Assistant/src/csharp/skills/calendarskill/calendarskill/Dialogs/UpdateEvent/UpdateEventDialog.cs
@@ -6,6 +6,7 @@ using System.Threading.Tasks;
 using CalendarSkill.Common;
 using CalendarSkill.Dialogs.Main.Resources;
 using CalendarSkill.Dialogs.Shared;
+using CalendarSkill.Dialogs.Shared.DialogOptions;
 using CalendarSkill.Dialogs.Shared.Prompts.Options;
 using CalendarSkill.Dialogs.Shared.Resources;
 using CalendarSkill.Dialogs.UpdateEvent.Resources;
