+++ solutions/Virtual-Assistant/src/csharp/skills/calendarskill/calendarskill/Util/DisplayHelper.cs
@@ -1,8 +1,7 @@
-﻿using Microsoft.Bot.Solutions.Resources;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Threading.Tasks;
+using Microsoft.Bot.Solutions.Resources;
 using static CalendarSkill.Models.EventModel;
 
 namespace CalendarSkill.Util
